pyproject.toml
[project]
name = "INMA backend"
version = "0.1.0"
description = "INMA Backend FastAPI"
authors = [
  { name = "David Baños Abril"}
]
license = { text = "MIT" }

requires-python = ">=3.12"

dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "beanie>=2.0.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.0",
    "fastapi-sessions>=0.3.2",
    "ftfy>=6.3.1",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-openai>=1.0.1",
    "motor>=3.7.1",
    "numpy>=2.3.4",
    "openai>=2.6.0",
    "opencage>=3.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=7.0.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "sqlglot>=27.28.1",
    "twilio>=9.8.4",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.23.6",
  "httpx>=0.27.0",
  "ruff>=0.6.8",
  "black>=24.8.0",
  "deptry>=0.23.1"
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["uv"]
build-backend = "uv.build"
